requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
pypdf
faiss-cpu
groq
langchain-groq
langchainhub
PyPDF2
langchain-objectbox